chore: move grid dependent tests
