+++ powersimdata/input/tests/test_grid.py
@@ -7,33 +7,6 @@ import pytest
 from powersimdata.input.grid import Grid
 from powersimdata.input.helpers import add_column_to_data_frame
 from powersimdata.input.scenario_grid import format_gencost, link
-from powersimdata.network.usa_tamu.model import TAMU, check_interconnect
-
-
-def test_interconnect_type():
-    interconnect = "Western"
-    with pytest.raises(TypeError):
-        check_interconnect(interconnect)
-
-
-def test_interconnect_value():
-    interconnect = ["Canada"]
-    with pytest.raises(ValueError):
-        check_interconnect(interconnect)
-
-
-def test_interconnect_duplicate_value():
-    interconnect = ["Western", "Western", "Texas"]
-    with pytest.raises(
-        ValueError, match="List of interconnects contains duplicate values"
-    ):
-        check_interconnect(interconnect)
-
-
-def test_interconnect_usa_is_unique():
-    interconnect = ["Western", "USA"]
-    with pytest.raises(ValueError, match="USA interconnect cannot be paired"):
-        check_interconnect(interconnect)
 
 
 def test_add_column_to_data_frame():
@@ -47,35 +20,6 @@ def test_add_column_to_data_frame():
 def test_grid_type():
     g = Grid(["USA"])
     assert isinstance(g, Grid)
-
-
-def test_drop_one_interconnect():
-    model = TAMU(["Western", "Texas"])
-    assert model.interconnect == ["Western", "Texas"]
-    assert "eastern" not in model.sub.interconnect.unique()
-    assert "eastern" not in model.bus2sub.interconnect.unique()
-    assert "eastern" not in model.bus.interconnect.unique()
-    assert "eastern" not in model.plant.interconnect.unique()
-    assert "eastern" not in model.branch.interconnect.unique()
-    assert "eastern" not in model.gencost["before"].interconnect.unique()
-    assert "eastern" not in model.gencost["after"].interconnect.unique()
-    assert "eastern" not in model.dcline.from_interconnect.unique()
-    assert "eastern" not in model.dcline.to_interconnect.unique()
-
-
-def test_drop_two_interconnect():
-    model = TAMU(["Western"])
-    assert model.interconnect == ["Western"]
-    for interconnect in ["Eastern", "Texas"]:
-        assert interconnect not in model.sub.interconnect.unique()
-        assert interconnect not in model.bus2sub.interconnect.unique()
-        assert interconnect not in model.bus.interconnect.unique()
-        assert interconnect not in model.plant.interconnect.unique()
-        assert interconnect not in model.branch.interconnect.unique()
-        assert interconnect not in model.gencost["before"].interconnect.unique()
-        assert interconnect not in model.gencost["after"].interconnect.unique()
-        assert interconnect not in model.dcline.from_interconnect.unique()
-        assert interconnect not in model.dcline.to_interconnect.unique()
 
 
 def test_format_gencost_polynomial_only_same_n():
